Feature/CountDown

+++ server/routes/api/index.js
@@ -222,7 +222,7 @@ router.post('/profile/upload/image/:id', upload.single('profileImage'), async (r
   var date = new Date().toISOString().replace(/T/, ' ').replace(/\..+/, '').replace(' ', '').replace(':', '').replace(':', '').replace('-', '').replace('-', '')
   if (user = await users.findOne({ "username": req.params.id, "isConfirmed": true})) {
     req.file.path = req.file.path.replace(/\\/g, "/")
-    if(typeof user.profileImage !== 'undefined' || user.profileImage != null){
+    if((typeof user.profileImage !== 'undefined' && user.profileImage != null)){
       if (fs.existsSync('static' + user.profileImage.path)){
         fs.unlinkSync('static' + user.profileImage.path);
       }
